requirements.txt
numpy>=1.22
scipy>=1.8
pandas>=1.4
scikit-learn>=1.2
umap-learn>=0.5
matplotlib>=3.7
